% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documented_objects.R
\docType{data}
\name{arabidopsis_chromosomes}
\alias{arabidopsis_chromosomes}
\title{Chromosome lengths for Arabidopsis}
\format{A data.frame.}
\description{
A data.frame with chromosome lengths for Arabidopsis.
}
\examples{
data(arabidopsis_chromosomes)
print(arabidopsis_chromosomes)
}
