% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documented_objects.R
\docType{data}
\name{arabidopsis_genes}
\alias{arabidopsis_genes}
\title{Gene coordinates for Arabidopsis (chr1)}
\format{A \code{\link[GenomicRanges]{GRanges}} object.}
\description{
A \code{\link[GenomicRanges]{GRanges}} object for demonstration purposes in examples of package \pkg{\link{methimpute}}. The object contains gene coordinates of chr1 from Arabidopsis.
}
\examples{
data(arabidopsis_genes)
print(arabidopsis_genes)
}

