% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documented_objects.R
\docType{data}
\name{arabidopsis_TEs}
\alias{arabidopsis_TEs}
\title{Transposable element coordinates for arabidopsis (chr1)}
\format{A \code{\link[GenomicRanges]{GRanges}} object.}
\description{
A \code{\link[GenomicRanges]{GRanges}} object for demonstration purposes in examples of package \pkg{\link{methimpute}}. The object contains transposable element coordinates of chr1 from arabidopsis.
}
\examples{
data(arabidopsis_TEs)
print(arabidopsis_TEs)
}

