% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binMethylome.R
\name{binCounts}
\alias{binCounts}
\title{Bin counts in windows}
\usage{
binCounts(data, binsize)
}
\arguments{
\item{data}{A \code{\link[GenomicRanges]{GRanges}} object with metadata column 'counts' which is a matrix with columns 'methylated' and 'total'.}

\item{binsize}{The window size used for binning.}
}
\value{
A \code{\link[GenomicRanges]{GRanges}} object.
}
\description{
Bin counts from cytosines in equidistant bins.
}

