% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binMethylome.R
\name{binMethylome}
\alias{binMethylome}
\title{Bin cytosine positions in windows}
\usage{
binMethylome(data, binsize)
}
\arguments{
\item{data}{A \code{\link[GenomicRanges]{GRanges}} object with metadata columns 'methylated', 'context', 'counts.unmethylated' and 'counts.methylated'.}

\item{binsize}{The window size used for binning.}
}
\value{
A list() with \code{\link[GenomicRanges]{GRanges}} objects.
}
\description{
Bin cytosine positions, methylation counts and state in equidistant windows.
}

