% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binMethylome.R
\name{binPositions}
\alias{binPositions}
\title{Bin positions in windows}
\usage{
binPositions(data, binsize)
}
\arguments{
\item{data}{A \code{\link[GenomicRanges]{GRanges}} object with metadata column 'context'.}

\item{binsize}{The window size used for binning.}
}
\value{
A \code{\link[GenomicRanges]{GRanges}} object.
}
\description{
Bin cytosine positions in equidistant bins.
}
