% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callMethylation.R
\name{callMethylation}
\alias{callMethylation}
\title{Call methylated regions}
\usage{
callMethylation(data, per.chrom = FALSE, num.threads = 1,
  fit.on.chrom = "chr1", transDist = 10000, eps = 0.01, max.time = Inf,
  max.iter = Inf, count.cutoff = 1000, verbosity = 1)
}
\arguments{
\item{data}{A \code{\link[GenomicRanges]{GRanges}} object with metadata columns 'distance' and 'counts'.}

\item{per.chrom}{Parallelize the fitting of the multivariate HMM per chromosome using \code{num.threads}.}

\item{num.threads}{Number of threads to use if \code{per.chrom = TRUE}.}

\item{fit.on.chrom}{A character vector giving the chromosomes on which the HMM will be fitted.}

\item{transDist}{The exponential decaying constant for the distance-dependent transition matrix. Should be given in the same units as \code{distances}.}

\item{eps}{Convergence threshold for the Baum-Welch algorithm.}

\item{max.time}{Maximum running time in seconds for the Baum-Welch algorithm.}

\item{max.iter}{Maximum number of iterations for the Baum-Welch algorithm.}

\item{count.cutoff}{\code{counts} > \code{count.cutoff} will be set to \code{count.cutoff}. The purpose is to increase the speed of the Baum-Welch algorithm, which is slowed down by huge values in \code{counts}.}

\item{verbosity}{Integer from c(0,1) specifying the verbosity of the fitting procedure.}
}
\value{
A list with fitted parameters, posteriors and input parameters.
}
\description{
Call methylated, unmethylated and hemimethylated regions by fitting a multivariate Hidden Markov Model to the data.
}
\details{
In a first step, the function will use \code{\link{fitSignalBackground}} to obtain the distribution parameters for a signal-background model for methylated and unmethylated read counts, respectively. The resulting distributions are used to fit a multivariate HMM with states \code{c("Background", "Methylated", "UNmethylated", "Hemimethylated")}.
}

