% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callMethylation.R
\name{callMethylationBinomial}
\alias{callMethylationBinomial}
\title{Methylation calling with a binomial-test HMM}
\usage{
callMethylationBinomial(data, fit.on.chrom = NULL, transDist = 10000,
  eps = 0.01, max.time = Inf, max.iter = Inf, quantile.cutoff = 1,
  verbosity = 1, initial.params = NULL, include.heterozygosity = FALSE)
}
\value{
A list with fitted parameters, posteriors.
}
\description{
Call methylation status of cytosines (or bins) with a Hidden Markov Model using a binomial test for the emission probabilities.
}

