% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dbetaMirror}
\alias{dbetaMirror}
\title{Beta mirror distribution}
\usage{
dbetaMirror(x, emissionParams, weights)
}
\arguments{
\item{emissionParams}{A matrix with emission parameters.}

\item{weights}{A table giving the weights for the two components.}
}
\description{
Get values for the beta mirror distribution
}

