% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dbetaSymmetric}
\alias{dbetaSymmetric}
\title{Beta symmetric distribution}
\usage{
dbetaSymmetric(x, emissionParams, weights)
}
\arguments{
\item{x}{A vector of quantiles.}

\item{emissionParams}{A matrix with emission parameters.}

\item{weights}{A table giving the weights for the two components.}
}
\description{
Get values for the beta symmetric distribution
}

