% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmentAnalysis.R
\name{enrichmentAtAnnotation}
\alias{enrichmentAtAnnotation}
\title{Enrichment of (combinatorial) states for genomic annotations}
\usage{
enrichmentAtAnnotation(bins, info, annotation, bp.around.annotation = 10000,
  region = c("start", "inside", "end"), what = c("combinations", "peaks",
  "counts"), num.intervals = 21, statistic = "fold")
}
\arguments{
\item{bins}{The \code{$bins} entry from a \code{\link{multimodel}} or \code{\link{combinedMultimodel}} object.}

\item{info}{The \code{$info} entry from a \code{\link{multimodel}} or \code{\link{combinedMultimodel}} object.}

\item{annotation}{A \code{\link{GRanges}} object with the annotation of interest.}

\item{bp.around.annotation}{An integer specifying the number of basepairs up- and downstream of the annotation for which the enrichment will be calculated.}

\item{region}{A combination of \code{c('start','inside','end')} specifying the region of the annotation for which the enrichment will be calculated. Select \code{'start'} if you have a point-sized annotation like transcription start sites. Select \code{c('start','inside','end')} if you have long annotations like genes.}

\item{what}{One of \code{c('combinations','peaks','counts','transitions')} specifying which statistic to calculate.}

\item{num.intervals}{Number of intervals for enrichment 'inside' of annotation.}

\item{statistic}{The statistic to calculate. Either 'fold' for fold enrichments or 'fraction' for fraction of bins falling into the annotation.}
}
\value{
A \code{list()} containing \code{data.frame()}s for enrichment of combinatorial states and binary states at the start, end and inside of the annotation.
}
\description{
The function calculates the enrichment of a genomic feature with peaks or combinatorial states. Input is a \code{\link{multimodel}} object (containing the peak calls and combinatorial states) and a \code{\link{GRanges}} object containing the annotation of interest (e.g. transcription start sites or genes).
}
\author{
Aaron Taudt
}

