% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callMethylation.R
\name{fitNComponentHMM}
\alias{fitNComponentHMM}
\title{Fit an n-component HMM}
\usage{
fitNComponentHMM(data, states = 0:5, observable = "counts",
  fit.on.chrom = NULL, transDist = 700, eps = 0.01, max.time = Inf,
  max.iter = Inf, count.cutoff = 500, verbosity = 1, num.threads = 1,
  initial.params = NULL)
}
\arguments{
\item{data}{A \code{\link[GenomicRanges]{GRanges}} object with metadata columns 'distance' and 'counts' (or any other column specified as \code{observable}).}

\item{states}{An integer vector giving the states for the Hidden Markov Model. State '0' will be modeled by a delta distribution, all other states ('1','2','3',...) with negative binomial distributions.}

\item{observable}{A character naming the column of \code{data} that will serve as observable for the HMM.}

\item{fit.on.chrom}{A character vector giving the chromosomes on which the HMM will be fitted.}

\item{transDist}{The exponential decaying constant for the distance-dependent transition matrix. Should be given in the same units as \code{distances}.}

\item{eps}{Convergence threshold for the Baum-Welch algorithm.}

\item{max.time}{Maximum running time in seconds for the Baum-Welch algorithm.}

\item{max.iter}{Maximum number of iterations for the Baum-Welch algorithm.}

\item{count.cutoff}{A cutoff for the \code{observable}. Lower cutoffs will speed up the fitting procedure and improve convergence in some cases. Set to \code{Inf} to disable this filtering.}

\item{verbosity}{Integer from c(0,1) specifying the verbosity of the fitting procedure.}

\item{num.threads}{Number of CPU to use for the computation.}

\item{initial.params}{An \code{\link{NcomponentHMM}} or a file that contains such an object. Parameters from this model will be used for initialization of the fitting procedure.}
}
\value{
A list with fitted parameters, posteriors, and the input parameters.
}
\description{
Fit an n-component Hidden Markov Model to the supplied counts. The transition matrix is distance-dependent with exponential decaying constant \code{transDist} (only relevant in non-consecutive bins). The zero-th component is a delta distribution to account for empty bins, and all other n-components are modeled as negative binomial distributions.
}

