% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callMethylation.R
\name{fitRatio}
\alias{fitRatio}
\title{Fit a three-component HMM for methylation levels}
\usage{
fitRatio(data, fit.on.chrom = NULL, transDist = 10000, eps = 0.01,
  max.time = Inf, max.iter = Inf, verbosity = 1)
}
\arguments{
\item{data}{A \code{\link[GenomicRanges]{GRanges}} object with metadata columns 'distance' and 'ratio'.}

\item{fit.on.chrom}{A character vector giving the chromosomes on which the HMM will be fitted.}

\item{transDist}{The exponential decaying constant for the distance-dependent transition matrix. Should be given in the same units as \code{distances}.}

\item{eps}{Convergence threshold for the Baum-Welch algorithm.}

\item{max.time}{Maximum running time in seconds for the Baum-Welch algorithm.}

\item{max.iter}{Maximum number of iterations for the Baum-Welch algorithm.}

\item{verbosity}{Integer from c(0,1) specifying the verbosity of the fitting procedure.}
}
\value{
A list with fitted parameters, posteriors, and the input parameters.
}
\description{
Fit a three-component Hidden Markov Model to the supplied counts. The transition matrix is distance-dependent with exponential decaying constant \code{transDist}. Components are modeled as beta distributions.
}

