% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPosteriors.R
\name{getPosteriors}
\alias{getPosteriors}
\title{Get original posteriors}
\usage{
getPosteriors(data)
}
\arguments{
\item{data}{The \code{$data} entry from a \code{\link{methimputeBinomialHMM}} object.}
}
\value{
A matrix with posteriors.
}
\description{
Transform the 'posteriorMeth', 'posteriorMax', and 'status' columns into original posteriors from the HMM.
}
