% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{heatmapTransitionProbs}
\alias{heatmapTransitionProbs}
\title{Heatmap of transition probabilities}
\usage{
heatmapTransitionProbs(model, order = FALSE)
}
\arguments{
\item{model}{A \code{\link{multiHMM}} object or file that contains such an object.}

\item{order}{Set to \code{TRUE} if you want to order the heatmap.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
}
\description{
Plot a heatmap of transition probabilities for a \code{\link{multiHMM}} model.
}
\seealso{
\code{\link{plotting}}
}

