% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{import}
\alias{import}
\alias{importBSSeeker}
\alias{importBismark}
\alias{importMethylpy}
\title{Methimpute data import}
\usage{
importMethylpy(file, chrom.lengths = NULL)

importBSSeeker(file, chrom.lengths = NULL)

importBismark(file, chrom.lengths = NULL)
}
\arguments{
\item{file}{The file to import.}

\item{chrom.lengths}{A data.frame with chromosome names in the first, and chromosome lengths in the second column. Only chromosomes named in here will be returned. Alternatively a tab-separated file with such a data.frame (with headers).}
}
\value{
A \code{\link{methimputeData}} object.
}
\description{
This page provides an overview of all \pkg{\link{methimpute}} data import functions.
}
\section{Functions}{
\itemize{
\item \code{importMethylpy}: Import a Methylpy methylation extractor file.

\item \code{importBSSeeker}: Import a BSSeeker methylation extractor file.

\item \code{importBismark}: Import a Bismark methylation extractor file.
}}
\examples{
## Get an example file in BSSeeker format
file <- system.file("extdata","arabidopsis_bsseeker.txt.gz", package="methimpute")
data(arabidopsis_chromosomes)
bsseeker.data <- importBSSeeker(file, chrom.lengths=arabidopsis_chromosomes)

## Get an example file in Bismark format
file <- system.file("extdata","arabidopsis_bismark.txt", package="methimpute")
data(arabidopsis_chromosomes)
arabidopsis_chromosomes$chromosome <- sub('chr', '', arabidopsis_chromosomes$chromosome)
bismark.data <- importBismark(file, chrom.lengths=arabidopsis_chromosomes)

}

