% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{importBismarck}
\alias{importBismarck}
\title{Import a Bismarck methylation extractor file}
\usage{
importBismarck(files, chrom.lengths = NULL,
  temp.store = tempfile("importBismarck"))
}
\arguments{
\item{files}{The files to import.}

\item{chrom.lengths}{A named vector containing the chromosome lengths. Only chromosomes named in here will be returned.}

\item{temp.store}{A folder where to save temporary files on disk. Set to \code{NULL} if no temporary files should be created.}
}
\value{
A \code{\link[GenomicRanges]{GRanges}} object with metadata columns 'methylated' and 'context'.
}
\description{
Import a Bismarck methylation extractor file into a \code{\link[GenomicRanges]{GRanges}} object.
}

