% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{importBismarck_old}
\alias{importBismarck_old}
\title{Import a Bismarck methylation extractor file}
\usage{
importBismarck_old(files, chrom.lengths.df, temp.store = NULL)
}
\arguments{
\item{files}{The files to import.}

\item{chrom.lengths.df}{A data.frame obtained with \code{\link[GenomeInfoDb]{fetchExtendedChromInfoFromUCSC}} for the seqlengths.}

\item{temp.store}{A folder where to save temporary files on disk. Set to \code{TRUE} if memory is limiting.}
}
\value{
A \code{\link[GenomicRanges]{GRanges}} object with metadata columns 'methylated' and 'context'.
}
\description{
Import a Bismarck methylation extractor file into a \code{\link[GenomicRanges]{GRanges}} object.
}

