% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{importBismark}
\alias{importBismark}
\title{Import a Bismark methylation extractor file}
\usage{
importBismark(files, chrom.lengths = NULL,
  temp.store = tempfile("importBismark"))
}
\arguments{
\item{files}{The files to import.}

\item{chrom.lengths}{A named vector containing the chromosome lengths. Only chromosomes named in here will be returned.}

\item{temp.store}{A folder where to save temporary files on disk. Set to \code{NULL} if no temporary files should be created.}
}
\value{
A \code{\link{methimputeData}} object.
}
\description{
Import a Bismark methylation extractor file into a \code{\link[GenomicRanges]{GRanges}} object.
}

