% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{importBSSeeker}
\alias{importBSSeeker}
\title{Import a BSSeeker methylation extractor file}
\usage{
importBSSeeker(file, chrom.lengths = NULL)
}
\arguments{
\item{file}{The file to import.}

\item{chrom.lengths}{A named vector containing the chromosome lengths. Only chromosomes named in here will be returned.}
}
\value{
A \code{\link{methimputeData}} object.
}
\description{
Import a BSSeeker methylation extractor file into a \code{\link[GenomicRanges]{GRanges}} object.
}
\examples{
## Get an example file in BSSeeker format
file <- system.file("extdata","arabidopsis_bsseeker.txt.gz", package="methimpute")
data(arabidopsis_chromosomes)
bsseeker.data <- importBSSeeker(file, chrom.lengths=arabidopsis_chromosomes)

}

