% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{importMethylpy}
\alias{importMethylpy}
\title{Import a Methylpy methylation extractor file}
\usage{
importMethylpy(file, chrom.lengths = NULL)
}
\arguments{
\item{file}{The file to import.}

\item{chrom.lengths}{A named vector containing the chromosome lengths. Only chromosomes named in here will be returned.}
}
\value{
A \code{\link{methimputeData}} object.
}
\description{
Import a Methylpy methylation extractor file into a \code{\link[GenomicRanges]{GRanges}} object.
}

