% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeConsensus.R
\name{makeConsensus}
\alias{makeConsensus}
\title{Make consensus segments}
\usage{
makeConsensus(models)
}
\description{
Given a list of \code{\link[GenomicRanges]{GRanges}}, construct a new GRanges as a \code{\link[GenomicRanges]{disjoin}} of the input segments.
}

