% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documented_objects.R
\name{methimpute-objects}
\alias{methimpute-objects}
\title{methimpute objects}
\description{
\pkg{\link{methimpute}} defines several objects.
\itemize{
\item \code{\link{methimputeData}}: Returned by \code{\link{importBSSeeker}}, \code{\link{importBismarck}} and \code{\link{inflateMethylome}}.
\item \code{\link{methimputeBinomialHMM}}: Returned by \code{\link{callMethylation}}.
}
}

