% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methimpute-package.R
\docType{package}
\name{methimpute-package}
\alias{methimpute}
\alias{methimpute-package}
\title{methIMPUTE: Imputation-guided methylation status calling for WGBS-seq}
\description{
\pkg{methimpute} is an R-package for methylation status calling in Whole-Genome Bisulfite-sequencing (WGBS-seq) data. Its powerful Hidden Markov model implementation enables imputation of methylation status calls for cytosines without any coverage.
}
\details{
Please read the vignette for a tutorial on how to use this package. You can do this by typing \code{browseVignettes("methimpute")}. Here is an overview of all \code{\link{plotting}} functions.
}
\author{
Aaron Taudt
}

