% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documented_objects.R
\name{methimputeData}
\alias{methimputeData}
\title{methimputeData}
\description{
A \code{\link[GenomicRanges]{GRanges}} object containing cytosine coordinates with meta-data columns 'context' and 'counts'.
}
\seealso{
\code{\link{methimpute-objects}}
}

