% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callMethylation.R
\name{multivariateSegmentation}
\alias{multivariateSegmentation}
\title{Make a multivariate segmentation}
\usage{
multivariateSegmentation(models, ID, fit.on.chrom = NULL, transDist = 700,
  eps = 0.01, max.time = Inf, max.iter = Inf, max.states = Inf,
  verbosity = 1)
}
\value{
A list with fitted parameters, posteriors.
}
\description{
Make a multivariate segmentation by fitting the transition probabilities of a multivariate Hidden Markov Model.
}
\details{
The function will use the provided univariate Hidden Markov Models (HMMs) to build the multivariate emission density. Number of states are also taken from the combination of univaraite HMMs.
}

