% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotConvergence}
\alias{plotConvergence}
\title{Plot convergence info}
\usage{
plotConvergence(model)
}
\arguments{
\item{model}{A \code{\link{BinomialHMM}} object or file that contains such an object.}
}
\value{
A \code{\link[ggplot2]{ggplot}}.
}
\description{
Plot the convergence of the probability parameter in different sequence contexts.
}

