% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotEnrichment}
\alias{plotEnrichment}
\title{Plot enrichment profile}
\usage{
plotEnrichment(data, annotation, windowsize = 100, insidewindows = 20,
  range = 1000, category.column = NULL)
}
\arguments{
\item{data}{A \code{\link{methimputeData}} object or the \code{$data} entry of a \code{\link{methimputeBinomialHMM}} object.}

\item{annotation}{A \code{\link[GenomicRanges]{GRanges}} object with coordinates for the annotation.}

\item{windowsize}{Resolution in base-pairs for the curve upstream and downstream of the annotation.}

\item{insidewindows}{Number of data points for the curve inside the annotation.}

\item{range}{Distance upstream and downstream for which the enrichment profile is calculated.}

\item{category.column}{The name of a column in \code{data} that will be used for facetting of the plot.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
}
\description{
Plot an enrichment profile around an annotation.
}
\seealso{
\code{\link{plotting}}
}

