% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotHistogramBinomial}
\alias{plotHistogramBinomial}
\title{Plot a count histogram}
\usage{
plotHistogramBinomial(model, total.counts, binwidth = 1)
}
\description{
Plot a histogram of count values and fitted distributions.
}

