% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotHistogramRatio}
\alias{plotHistogramRatio}
\title{Plot a ratio histogram}
\usage{
plotHistogramRatio(hmm, num.intervals = 20)
}
\description{
Plot a histogram of ratio values and fitted distributions.
}

