% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotRatioHistogram}
\alias{plotRatioHistogram}
\title{Plot a ratio histogram}
\usage{
plotRatioHistogram(model, num.intervals = 20)
}
\description{
Plot a histogram of ratio values and fitted distributions.
}

