% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotScatter}
\alias{plotScatter}
\title{Scatter plot of methylation status calls}
\usage{
plotScatter(model, datapoints = 1000)
}
\arguments{
\item{model}{A \code{\link{methimputeBinomialHMM}} object.}

\item{datapoints}{The number of randomly selected datapoints for the plot.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
}
\description{
Plot a scatter plot of read counts colored by methylation status
}
\seealso{
\code{\link{plotting}}
}

