% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotStateHistograms}
\alias{plotStateHistograms}
\title{Histograms faceted by state}
\usage{
plotStateHistograms(segmentation, bins, plotcol, binwidth)
}
\arguments{
\item{segmentation}{A segmentation with metadata column 'state'.}

\item{bins}{A \code{\link{binnedMethylome}}.}

\item{plotcol}{A character vector specifying the meta-data columns for which to produce the faceted plot.}

\item{binwidth}{Bin width for the histogram.}
}
\value{
A \code{\link[ggplot2]{ggplot}}.
}
\description{
Make histograms of variables in \code{bins} faceted by state in \code{segmentation}.
}

