% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotStateScatter}
\alias{plotStateScatter}
\title{Scatter plots faceted by state}
\usage{
plotStateScatter(segmentation, bins, x, y, col = NULL, datapoints = 1000)
}
\arguments{
\item{segmentation}{A segmentation with metadata column 'state'.}

\item{bins}{A \code{\link{binnedMethylome}}.}

\item{x}{A character specifying the meta-data column that should go on the x-axis.}

\item{y}{A character specifying the meta-data column that should go on the y-axis.}

\item{col}{A character specifying the meta-data column that should be used for coloring.}
}
\value{
A \code{\link[ggplot2]{ggplot}}.
}
\description{
Make scatter plots of two variables in \code{bins} faceted by state in \code{segmentation}.
}

