\name{popmeth-package}
\alias{popmeth-package}
\alias{popmeth}
\docType{package}
\title{
\packageTitle{popmeth}
}
\description{
\packageDescription{popmeth}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{popmeth}
\packageIndices{popmeth}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{popmeth}

Maintainer: \packageMaintainer{popmeth}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
