% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printMethods.R
\name{print.BinomialHMMcontext}
\alias{print.BinomialHMMcontext}
\title{Print model object}
\usage{
\method{print}{BinomialHMMcontext}(model, ...)
}
\arguments{
\item{model}{A \code{\link{BinomialHMMcontext}} object.}

\item{...}{Ignored.}
}
\description{
Print model object
}

