% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printMethods.R
\name{print.methimputeBinomialHMM}
\alias{print.methimputeBinomialHMM}
\title{Print model object}
\usage{
\method{print}{methimputeBinomialHMM}(x, ...)
}
\arguments{
\item{x}{A \code{\link{methimputeBinomialHMM}} object.}

\item{...}{Ignored.}
}
\value{
An invisible \code{NULL}.
}
\description{
Print model object
}
