% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binMethylome.R
\name{binMethylome}
\alias{binMethylome}
\title{Bin cytosine positions in windows}
\usage{
binMethylome(data, binsize, contexts = "total",
  columns.average = c("posteriorMeth"))
}
\arguments{
\item{data}{A \code{\link[GenomicRanges]{GRanges}} object with metadata columns 'methylated', 'context', 'counts.unmethylated' and 'counts.methylated'.}

\item{binsize}{The window size used for binning.}

\item{contexts}{A character vector with contexts for which the binning will be done.}

\item{columns.average}{A character vector with names of columns in \code{data} that should be averaged in bins.}
}
\value{
A list() with a \code{\link[GenomicRanges]{GRanges}} objects for each context.
}
\description{
Bin cytosine positions, methylation counts and state in equidistant windows.
}
