% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callMethylation.R
\name{callMethylation}
\alias{callMethylation}
\title{Call methylated regions}
\usage{
callMethylation(data, fit.on.chrom = NULL, transDist = 10000, eps = 0.01,
  max.time = Inf, max.iter = Inf, quantile.cutoff = 0.999,
  states = c("Background", "Methylated", "UNmethylated", "Hemimethylated"),
  verbosity = 1)
}
\arguments{
\item{data}{A \code{\link[GenomicRanges]{GRanges}} object with metadata columns 'distance' and 'counts'.}

\item{fit.on.chrom}{A character vector giving the chromosomes on which the HMM will be fitted.}

\item{transDist}{The exponential decaying constant for the distance-dependent transition matrix. Should be given in the same units as \code{distances}.}

\item{eps}{Convergence threshold for the Baum-Welch algorithm.}

\item{max.time}{Maximum running time in seconds for the Baum-Welch algorithm.}

\item{max.iter}{Maximum number of iterations for the Baum-Welch algorithm.}

\item{quantile.cutoff}{A quantile (between 0 and 1) serving as cutoff for the \code{observable}. Lower cutoffs will speed up the fitting procedure and improve convergence in some cases. Set to 1 to disable this filtering.}

\item{verbosity}{Integer from c(0,1) specifying the verbosity of the fitting procedure.}
}
\value{
A list with fitted parameters, posteriors and input parameters.
}
\description{
Call methylated, unmethylated and hemimethylated regions by fitting a multivariate Hidden Markov Model to the data.
}
\details{
In a first step, the function will use \code{\link{fitSignalBackground}} to obtain the distribution parameters for a signal-background model for methylated and unmethylated read counts, respectively. The resulting distributions are used in a second step to fit a multivariate HMM with states \code{c("Background", "Methylated", "UNmethylated", "Hemimethylated")}.
}

