% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callMethylation.R
\name{callMethylationBinomial}
\alias{callMethylationBinomial}
\title{Methylation calling with a binomial-test HMM}
\usage{
callMethylationBinomial(data, fit.on.chrom = NULL, min.reads = 0,
  transDist = 700, eps = 0.01, max.time = Inf, max.iter = Inf,
  count.cutoff = 500, verbosity = 1, num.threads = 1,
  initial.params = NULL, include.heterozygosity = TRUE)
}
\value{
A list with fitted parameters, posteriors.
}
\description{
Call methylation status of cytosines (or bins) with a Hidden Markov Model using a binomial test for the emission probabilities.
}

