% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{distanceCorrelation}
\alias{distanceCorrelation}
\title{Estimate the \code{transDist} parameter}
\usage{
distanceCorrelation(data, distances = 0:50)
}
\value{
A list() containing a data.frame and a ggplot.
}
\description{
Estimate the \code{transDist} parameter from data.
}

