% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{estimateTransDist}
\alias{estimateTransDist}
\title{Obtain \code{transDist} parameter.}
\usage{
estimateTransDist(distcor)
}
\arguments{
\item{distcor}{The list produced by \code{\link{distanceCorrelation}}.}
}
\description{
Obtain an estimate for the \code{transDist} parameter in function \code{\link{callMethylation}} by fitting an exponential to the supplied correlations.
}

