% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractCytosinesFromFASTA.R
\name{extractCytosinesFromFASTA}
\alias{extractCytosinesFromFASTA}
\title{Extract cytosine coordinates}
\usage{
extractCytosinesFromFASTA(file, contexts = c("CG", "CHG", "CHH"))
}
\arguments{
\item{file}{A character with the file name.}

\item{contexts}{The contexts that should be extracted.}
}
\description{
Extract cytosine coordinates and context information from a FASTA file. Cytosines in ambiguous contexts are not reported.
}
\examples{
## Read a non-compressed FASTA files:
filepath <- system.file("extdata", "someORF.fa", package="Biostrings")
cytosines <- extractCytosinesFromFASTA(filepath)
}

