% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callMethylation.R
\name{fitSignalBackground}
\alias{fitSignalBackground}
\title{Fit a two-component HMM}
\usage{
fitSignalBackground(data, observable = "counts", fit.on.chrom = "chr1",
  transDist = 10000, eps = 0.01, max.time = Inf, max.iter = Inf,
  count.cutoff = 1000, verbosity = 1)
}
\arguments{
\item{data}{A \code{\link[GenomicRanges]{GRanges}} object with metadata columns 'distance' and 'counts'.}

\item{observable}{A character naming the column of \code{data} that will serve as observable for the HMM.}

\item{fit.on.chrom}{A character vector giving the chromosomes on which the HMM will be fitted.}

\item{transDist}{The exponential decaying constant for the distance-dependent transition matrix. Should be given in the same units as \code{distances}.}

\item{eps}{Convergence threshold for the Baum-Welch algorithm.}

\item{max.time}{Maximum running time in seconds for the Baum-Welch algorithm.}

\item{max.iter}{Maximum number of iterations for the Baum-Welch algorithm.}

\item{count.cutoff}{\code{counts} > \code{count.cutoff} will be set to \code{count.cutoff}. The purpose is to increase the speed of the Baum-Welch algorithm, which is slowed down by huge values in \code{counts}.}

\item{verbosity}{Integer from c(0,1) specifying the verbosity of the fitting procedure.}
}
\value{
A list with fitted parameters, posteriors, and the input parameters.
}
\description{
Fit a two-component Hidden Markov Model to the supplied \code{counts}. The transition matrix is distance-dependent with exponential decaying constant \code{transDist}. Components are modeled as negative binomial distributions.
}

