% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{importBismark}
\alias{importBismark}
\title{Import a Bismark methylation extractor file}
\usage{
importBismark(file, chrom.lengths = NULL)
}
\arguments{
\item{file}{The file to import.}

\item{chrom.lengths}{A data.frame with chromosome names in the first, and chromosome lengths in the second column. Only chromosomes named in here will be returned. Alternatively a tab-separated file with such a data.frame (with headers).}
}
\value{
A \code{\link{methimputeData}} object.
}
\description{
Import a Bismark methylation extractor file into a \code{\link[GenomicRanges]{GRanges}} object.
}
\examples{
## Get an example file in BSSeeker format
file <- system.file("extdata","arabidopsis_bismark.txt", package="methimpute")
data(arabidopsis_chromosomes)
names(arabidopsis_chromosomes) <- sub('chr', '', names(arabidopsis_chromosomes))
bismark.data <- importBismark(file, chrom.lengths=arabidopsis_chromosomes)

}

