% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{importBSSeeker}
\alias{importBSSeeker}
\title{Import a BSSeeker methylation extractor file}
\usage{
importBSSeeker(file, chrom.lengths = NULL,
  temp.store = tempfile("importBSSeeker"))
}
\arguments{
\item{file}{The file to import.}

\item{chrom.lengths}{A named vector containing the chromosome lengths.}

\item{temp.store}{A folder where to save temporary files on disk. Set to \code{NULL} if no temporary files should be created.}
}
\value{
A \code{\link[GenomicRanges]{GRanges}} object with metadata columns 'methylated' and 'context'.
}
\description{
Import a BSSeeker methylation extractor file into a \code{\link[GenomicRanges]{GRanges}} object.
}

