% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{importMethylpy}
\alias{importMethylpy}
\title{Import a Methylpy methylation extractor file}
\usage{
importMethylpy(file, chrom.lengths = NULL)
}
\arguments{
\item{file}{The file to import.}

\item{chrom.lengths}{A data.frame with chromosome names in the first, and chromosome lengths in the second column. Only chromosomes named in here will be returned. Alternatively a tab-separated file with such a data.frame (with headers).}
}
\value{
A \code{\link{methimputeData}} object.
}
\description{
Import a Methylpy methylation extractor file into a \code{\link[GenomicRanges]{GRanges}} object.
}

