% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{importRene}
\alias{importRene}
\title{Import a Rene methylation extractor file}
\usage{
importRene(file, chrom.lengths = NULL)
}
\arguments{
\item{file}{The file to import.}

\item{chrom.lengths}{A named vector containing the chromosome lengths. Only chromosomes named in here will be returned.}
}
\value{
A \code{\link[GenomicRanges]{GRanges}} object with metadata columns 'methylated' and 'context'.
}
\description{
Import a Rene methylation extractor file into a \code{\link[GenomicRanges]{GRanges}} object.
}

