% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{importRene}
\alias{importRene}
\title{Import a Rene methylation extractor file}
\usage{
importRene(file, chrom.lengths.df = NULL,
  temp.store = tempfile("importRene"))
}
\arguments{
\item{file}{The file to import.}

\item{chrom.lengths.df}{A data.frame obtained with \code{\link[GenomeInfoDb]{fetchExtendedChromInfoFromUCSC}} for the seqlengths.}

\item{temp.store}{A folder where to save temporary files on disk. Set to \code{NULL} if no temporary files should be created.}
}
\value{
A \code{\link[GenomicRanges]{GRanges}} object with metadata columns 'methylated' and 'context'.
}
\description{
Import a Rene methylation extractor file into a \code{\link[GenomicRanges]{GRanges}} object.
}

