% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{inflateMethylome}
\alias{inflateMethylome}
\title{Inflate an imported methylation extractor file}
\usage{
inflateMethylome(methylome, methylome.full)
}
\arguments{
\item{methylome}{A \code{\link[GenomicRanges]{GRanges}} with methylation counts.}

\item{methylome.full}{A \code{\link[GenomicRanges]{GRanges}} with positions for all cytosines or a file with such an object.}
}
\value{
The \code{methylome.full} object with added metadata column 'counts'.
}
\description{
Inflate an imported methylation extractor file to more cytosines. This is useful to obtain a full methylome, including non-covered cytosines.
}

