% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documented_objects.R
\name{methimputeBinomialHMM}
\alias{methimputeBinomialHMM}
\title{methimputeBinomialHMM}
\value{
A list() with the following entries:
\item{convergenceInfo}{A list() with information about the convergence of the model fitting procedure.}
\item{params}{A list() with fitted and non-fitted model parameters.}
\item{params.initial}{A list() with initial values for the model parameters.}
\item{data}{A \code{\link[GenomicRanges]{GRanges}} with cytosine positions and methylation status calls.}
\item{segments}{The \code{data} entry where coordinates of consecutive cytosines with the same methylation status have been merged.}
}
\description{
The \code{methimputeBinomialHMM} is a list() which contains various entries (see Value section)
}
\seealso{
\code{\link{methimpute-objects}}
}

