% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotting}
\alias{plotConvergence}
\alias{plotEnrichment}
\alias{plotHistogram}
\alias{plotPosteriorDistance}
\alias{plotScatter}
\alias{plotTransitionProbs}
\alias{plotting}
\title{methimpute plotting functions}
\usage{
plotHistogram(model, total.counts, binwidth = 1)

plotScatter(model, datapoints = 1000)

plotTransitionProbs(model)

plotConvergence(model)

plotEnrichment(data, annotation, windowsize = 100, insidewindows = 20,
  range = 1000, category.column = NULL)

plotPosteriorDistance(data, datapoints = 1e+06, binwidth = 5,
  max.coverage.y = 0, min.coverage.x = 3, xmax = 200,
  xbreaks.interval = xmax/10)
}
\arguments{
\item{model}{A \code{\link{methimputeBinomialHMM}} object.}

\item{total.counts}{The number of total counts for which the histogram is to be plotted.}

\item{binwidth}{The bin width for the histogram/boxplot.}

\item{datapoints}{The number of randomly selected datapoints for the plot.}

\item{data}{A \code{\link{methimputeData}} object or the \code{$data} entry of a \code{\link{methimputeBinomialHMM}} object.}

\item{annotation}{A \code{\link[GenomicRanges]{GRanges}} object with coordinates for the annotation.}

\item{windowsize}{Resolution in base-pairs for the curve upstream and downstream of the annotation.}

\item{insidewindows}{Number of data points for the curve inside the annotation.}

\item{range}{Distance upstream and downstream for which the enrichment profile is calculated.}

\item{category.column}{The name of a column in \code{data} that will be used for facetting of the plot.}

\item{max.coverage.y}{Maximum coverage for cytosines on the y-axis.}

\item{min.coverage.x}{Minimum coverage for cytosines on the x-axis.}

\item{xmax}{Upper limit for the x-axis.}

\item{xbreaks.interval}{Interval for breaks on the x-axis.}

\item{data}{The \code{$data} entry of a \code{\link{methimputeBinomialHMM}} object.}

\item{datapoints}{The number of datapoints to consider for the plot.}

\item{binwidth}{The bin width for the histogram/boxplot.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
}
\description{
This page provides an overview of all \pkg{\link{methimpute}} plotting functions.
}
\section{Functions}{
\itemize{
\item \code{plotHistogram}: Plot a histogram of count values and fitted distributions.

\item \code{plotScatter}: Plot a scatter plot of read counts colored by methylation status.

\item \code{plotTransitionProbs}: Plot a heatmap of transition probabilities.

\item \code{plotConvergence}: Plot the convergence of the probability parameters.

\item \code{plotEnrichment}: Plot an enrichment profile around an annotation.

\item \code{plotPosteriorDistance}: Maximum posterior vs. distance to nearest covered cytosine.
}}
\examples{
## Get some toy data
file <- system.file("data","arabidopsis_toydata.RData",
                    package="methimpute")
data <- get(load(file))
print(data)
model <- callMethylation(data)
## Make nice plots
plotHistogram(model, total.counts=5)
plotScatter(model)
plotTransitionProbs(model)
plotConvergence(model)
plotPosteriorDistance(model$data)

## Get annotation data and make an enrichment profile
# Note that this looks a bit ugly because our toy data
# has only 200000 datapoints.
data(arabidopsis_genes)
plotEnrichment(model$data, annotation=arabidopsis_genes)

}

